requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.3.0=py38h06a4308_0
aiohttp=3.8.1=py38h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyh9f0ad1d_0
astor=0.8.1=py38h06a4308_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py38h06a4308_0
attrs=22.1.0=py38h06a4308_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
boost-cpp=1.74.0=h359cf19_6
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.9.24=ha878542_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
cryptography=38.0.1=py38h9ce1e76_0
cudatoolkit=11.1.1=h6406543_8
cudnn=8.1.0.77=h90431f1_0
dataclasses=0.8=pyh6d0b6a4_7
fftw=3.3.9=h27cfd23_1
flatbuffers=2.0.0=h2531618_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.2.0=py38h7f8727e_0
gast=0.4.0=pyh9f0ad1d_0
giflib=5.2.1=h7b6447c_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.46.1=h33aed49_0
grpcio=1.42.0=py38hce63b2e_0
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=h3ffc7dd_1
icu=69.1=h9c3ff4c_0
idna=3.4=py38h06a4308_0
importlib-metadata=4.11.3=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h7f8727e_0
keras=2.4.3=hd3eb1b0_0
keras-base=2.4.3=pyhd3eb1b0_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcurl=7.85.0=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.4.2=h295c915_4
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_2
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.3=h295c915_1
mako=1.2.4=pyhd8ed1ab_0
markdown=3.3.4=py38h06a4308_0
markupsafe=2.1.1=py38h0a891b7_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=py38h5eee18b_0
ncurses=6.3=h5eee18b_3
numpy=1.23.4=py38h14f4228_0
numpy-base=1.23.4=py38h31eccc5_0
oauthlib=3.2.1=py38h06a4308_0
openssl=1.1.1s=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
platformdirs=2.5.2=pyhd8ed1ab_1
protobuf=3.20.1=py38h295c915_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pycuda=2021.1=py38h36e3a89_3
pyjwt=2.4.0=py38h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.15=h3fd9d12_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytools=2022.1.13=pyhd8ed1ab_0
pyyaml=6.0=py38h7f8727e_1
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.28.1=py38h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.9.3=py38h14f4228_0
setuptools=65.5.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.39.3=h5082296_0
tensorboard=2.10.0=py38h06a4308_0
tensorboard-data-server=0.6.0=py38hca6d32c_0
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
tensorflow=2.4.1=mkl_py38hb2083e0_0
tensorflow-addons=0.13.0=py38_1
tensorflow-base=2.4.1=mkl_py38h43e0292_0
tensorflow-estimator=2.5.0=pyh7b7c402_0
tensorflow-gpu=2.5.0=py38_cuda11.1_cudnn8.1_1
termcolor=2.1.0=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
typeguard=2.13.0=py_1
typing-extensions=4.3.0=py38h06a4308_0
typing_extensions=4.3.0=py38h06a4308_0
urllib3=1.26.12=py38h06a4308_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=py38h5eee18b_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py38h5eee18b_0
zipp=3.8.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
